Factor out simpler components
